+++ client/components/Reports/index.js
@@ -1,0 +1,3 @@
+import Reports from './Reports';
+
+export default Reports;
