--- client/components/ReportsPage/index.js
@@ -1,3 +1,0 @@
-import ReportsPage from './ReportsPage';
-
-export default ReportsPage;
